requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.0.0
